[AMTH] Create new benchmarks file
As a first step, we only keep BV benchmarks.
Currently, we do not delete the old benchmarks, instead
we moved them to benchmarks/legacy in where they won't
be built by the Makefile.
